HAAR-833: Use get paged roles endpoint in external users api - instead of auth
