+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/AuthService.kt
@@ -14,7 +14,6 @@ import uk.gov.justice.digital.hmpps.manageusersapi.resource.Role
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.RoleAdminTypeAmendment
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.RoleDescriptionAmendment
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.RoleNameAmendment
-import uk.gov.justice.digital.hmpps.manageusersapi.resource.RolesPaged
 import java.time.Duration
 
 @Service
@@ -144,32 +143,6 @@ class AuthService(
 
   private fun Set<AdminType>.addDpsAdmTypeIfRequiredAsList() =
     (if (AdminType.DPS_LSA in this) (this + AdminType.DPS_ADM) else this).map { it.adminTypeCode }
-
-  fun getPagedRoles(
-    page: Int,
-    size: Int,
-    sort: String,
-    roleName: String?,
-    roleCode: String?,
-    adminTypes: List<AdminType>?
-  ): RolesPaged {
-
-    return authWebClient.get()
-      .uri { uriBuilder ->
-        uriBuilder
-          .path("/api/roles/paged")
-          .queryParam("page", page)
-          .queryParam("size", size)
-          .queryParam("sort", sort)
-          .queryParam("roleName", roleName)
-          .queryParam("roleCode", roleCode)
-          .queryParam("adminTypes", adminTypes)
-          .build()
-      }
-      .retrieve()
-      .bodyToMono(RolesPaged::class.java)
-      .block()!!
-  }
 
   fun getRoles(adminTypes: List<AdminType>?): List<Role> {
