+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/ExternalUsersApiService.kt
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.Role
+import uk.gov.justice.digital.hmpps.manageusersapi.resource.RolesPaged
 
 @Service
 class ExternalUsersApiService(@Qualifier("externalUsersWebClient") val externalUsersWebClient: WebClient) {
@@ -23,6 +24,32 @@ class ExternalUsersApiService(@Qualifier("externalUsersWebClient") val externalU
       }
       .retrieve()
       .bodyToMono(RoleList::class.java)
+      .block()!!
+  }
+
+  fun getPagedRoles(
+    page: Int,
+    size: Int,
+    sort: String,
+    roleName: String?,
+    roleCode: String?,
+    adminTypes: List<AdminType>?
+  ): RolesPaged {
+
+    return externalUsersWebClient.get()
+      .uri { uriBuilder ->
+        uriBuilder
+          .path("/roles/paged")
+          .queryParam("page", page)
+          .queryParam("size", size)
+          .queryParam("sort", sort)
+          .queryParam("roleName", roleName)
+          .queryParam("roleCode", roleCode)
+          .queryParam("adminTypes", adminTypes)
+          .build()
+      }
+      .retrieve()
+      .bodyToMono(RolesPaged::class.java)
       .block()!!
   }
 }
