+++ src/test/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/integration/wiremock/ExternalUsersApiMockServer.kt
@@ -176,6 +176,157 @@ class ExternalUsersApiMockServer : WireMockServer(WIREMOCK_PORT) {
     )
   }
 
+  fun stubGetAllRolesPage3Descending() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=3&size=4&sort=roleName,desc&roleName&roleCode&adminTypes"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              """{
+                      "content": [
+                          {
+                              "roleCode": "ROLE_8",
+                              "roleName": "role 8",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_7",
+                              "roleName": "role 7",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_6",
+                              "roleName": "role 6",
+                              "roleDescription": "Sixth role",
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_5",
+                              "roleName": "role 5",
+                              "roleDescription": "Fifth role",
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          }
+                      ],
+                      "pageable": {
+                          "sort": {
+                              "sorted": true,
+                              "unsorted": false,
+                              "empty": false
+                          },
+                          "offset": 12,
+                          "pageNumber": 3,
+                          "pageSize": 4,
+                          "paged": true,
+                          "unpaged": false
+                      },
+                      "last": false,
+                      "totalPages": 10,
+                      "totalElements": 37,
+                      "size": 4,
+                      "number": 3,
+                      "sort": {
+                          "sorted": true,
+                          "unsorted": false,
+                          "empty": false
+                      },
+                      "numberOfElements": 4,
+                      "first": false,
+                      "empty": false
+                  }
+              """.trimIndent()
+            )
+        )
+    )
+  }
+
+  fun stubGetAllRolesPagedFilterRoleCode() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=0&size=10&sort=roleName,asc&roleName&roleCode=account&adminTypes"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              apiResponseBodyPaged()
+            )
+        )
+    )
+  }
+
+  fun stubGetAllRolesPagedFilterRoleName() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=0&size=10&sort=roleName,asc&roleName=manager&roleCode&adminTypes"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              apiResponseBodyPaged()
+            )
+        )
+    )
+  }
+
+  fun stubGetAllRolesPagedFilterAdminType() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=0&size=10&sort=roleName,asc&roleName&roleCode&adminTypes=EXT_ADM"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              apiResponseBodyPaged()
+            )
+        )
+    )
+  }
+
+  fun stubGetAllRolesPagedFilterAdminTypes() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=0&size=10&sort=roleName,asc&roleName&roleCode&adminTypes=EXT_ADM&adminTypes=DPS_ADM"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              apiResponseBodyPaged()
+            )
+        )
+    )
+  }
+
+  fun stubGetAllRolesPagedUsingAllFilters() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=1&size=10&sort=roleName,asc&roleName=manager&roleCode=account&adminTypes=EXT_ADM&adminTypes=DPS_ADM"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              apiResponseBodyPaged()
+            )
+        )
+    )
+  }
+
   fun apiResponseBody() =
     """
       [
@@ -196,4 +347,202 @@ class ExternalUsersApiMockServer : WireMockServer(WIREMOCK_PORT) {
           }
       ]
     """.trimIndent()
+
+  fun stubGetAllRolesPaged() {
+    stubFor(
+      get(urlEqualTo("/roles/paged?page=0&size=10&sort=roleName,asc&roleName&roleCode&adminTypes"))
+        .willReturn(
+          aResponse()
+            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
+            .withBody(
+              """{
+                      "content": [
+                          {
+                              "roleCode": "AUDIT_VIEWER",
+                              "roleName": "Audit viewer",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "AUTH_GROUP_MANAGER",
+                              "roleName": "Auth Group Manager",
+                              "roleDescription": "Gives group manager ability to administer user in there groups",
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_1",
+                              "roleName": "role 1",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_2",
+                              "roleName": "role 2",
+                              "roleDescription": "Second role",
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_3",
+                              "roleName": "role 3",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_4",
+                              "roleName": "role 4",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_5",
+                              "roleName": "role 5",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_6",
+                              "roleName": "role 6",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_7",
+                              "roleName": "role 7",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          },
+                          {
+                              "roleCode": "ROLE_8",
+                              "roleName": "role 8",
+                              "roleDescription": null,
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  }
+                              ]
+                          }
+                      ],
+                      "pageable": {
+                          "sort": {
+                              "sorted": true,
+                              "unsorted": false,
+                              "empty": false
+                          },
+                          "offset": 0,
+                          "pageNumber": 0,
+                          "pageSize": 10,
+                          "paged": true,
+                          "unpaged": false
+                      },
+                      "last": false,
+                      "totalPages": 4,
+                      "totalElements": 37,
+                      "size": 10,
+                      "number": 0,
+                      "sort": {
+                          "sorted": true,
+                          "unsorted": false,
+                          "empty": false
+                      },
+                      "numberOfElements": 10,
+                      "first": true,
+                      "empty": false
+                  }
+              """.trimIndent()
+            )
+        )
+    )
+  }
+
+  fun apiResponseBodyPaged() = """{
+                      "content": [
+                          {
+                              "roleCode": "ACCOUNT_MANAGER",
+                              "roleName": "The group account manager",
+                              "roleDescription": "A group account manager - responsible for managing groups",
+                              "adminType": [
+                                  {
+                                      "adminTypeCode": "EXT_ADM",
+                                      "adminTypeName": "External Administrator"
+                                  },
+                                  {
+                                      "adminTypeCode": "DPS_ADM",
+                                      "adminTypeName": "DPS Central Administrator"
+                                  }
+                              ]
+                          }
+                      ],
+                      "pageable": {
+                          "sort": {
+                              "sorted": true,
+                              "unsorted": false,
+                              "empty": false
+                          },
+                          "offset": 0,
+                          "pageNumber": 1,
+                          "pageSize": 10,
+                          "paged": true,
+                          "unpaged": false
+                      },
+                      "last": false,
+                      "totalPages": 1,
+                      "totalElements": 1,
+                      "size": 10,
+                      "number": 1,
+                      "sort": {
+                          "sorted": true,
+                          "unsorted": false,
+                          "empty": false
+                      },
+                      "numberOfElements": 1,
+                      "first": false,
+                      "empty": false
+                  }
+  """.trimIndent()
 }
